Improve random number generation, Close #151

+++ src/DNTCaptcha.Core/ShowDigitsProvider.cs
@@ -30,7 +30,7 @@ public class ShowDigitsProvider : ICaptchaTextProvider
     /// <param name="number">input number</param>
     /// <param name="language">local language</param>
     /// <returns>the equivalent text</returns>
-    public string GetText(long number, Language language) =>
+    public string GetText(int number, Language language) =>
         _captchaOptions.AllowThousandsSeparators
             ? string.Format(CultureInfo.InvariantCulture, "{0:N0}", number)
             : number.ToString(CultureInfo.InvariantCulture);
